OK DAOS core

+++ src/main/java/daos/core/TicketDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import entities.core.Ticket;
 
-public interface TicketDao extends JpaRepository<Ticket, Integer> {
+public interface TicketDao extends JpaRepository<Ticket, Long> {
 
 }
